LayoutTests/imported/w3c/web-platform-tests/content-security-policy/gen/top.http-rp/script-src-self/worklet-audio.https-expected.txt

CONSOLE MESSAGE: Refused to load https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=keep-origin&action=purge&key=e7438aa8-40ec-4752-85f1-28ba17ee395f&path=%2Fmixed-content because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=no-redirect&action=purge&key=ecac46ab-1b07-4adc-9f0e-d1d8fcfb9534&path=%2Fmixed-content because it does not appear in the script-src directive of the Content Security Policy.
CONSOLE MESSAGE: Refused to load https://www1.localhost:9443/common/security-features/subresource/worker.py?redirection=swap-origin&action=purge&key=b0a5ac6d-dd12-4491-8950-cca0c00466a8&path=%2Fmixed-content because it does not appear in the script-src directive of the Content Security Policy.
Blocked access to external URL https://www1.localhost:9443/common/security-features/subresource/worker.py?action=purge&key=14fba811-5ddb-4194-8154-ef210608d575&path=%2Fmixed-content

PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and keep-origin redirection from https context.
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and keep-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and no-redirect redirection from https context.
PASS Content Security Policy: Expects allowed for worklet-audio to same-https origin and no-redirect redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and keep-origin redirection from https context.
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and keep-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and no-redirect redirection from https context.
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and no-redirect redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and swap-origin redirection from https context.
PASS Content Security Policy: Expects blocked for worklet-audio to cross-https origin and swap-origin redirection from https context.: securitypolicyviolation
PASS Content Security Policy: Expects blocked for worklet-audio to same-https origin and swap-origin redirection from https context.
FAIL Content Security Policy: Expects blocked for worklet-audio to same-https origin and swap-origin redirection from https context.: securitypolicyviolation assert_equals: One violation event should be fired expected 1 but got 0
